+++ examples/com/moilioncircle/examples/function/DumpFunctionExample.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2016-2017 Leon Chen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.moilioncircle.examples.function;
+
+import com.moilioncircle.redis.replicator.RedisReplicator;
+import com.moilioncircle.redis.replicator.Replicator;
+import com.moilioncircle.redis.replicator.event.Event;
+import com.moilioncircle.redis.replicator.event.EventListener;
+import com.moilioncircle.redis.replicator.rdb.dump.DumpRdbVisitor;
+import com.moilioncircle.redis.replicator.rdb.dump.datatype.DumpFunction;
+
+/**
+ * @author Leon Chen
+ * @since 3.6.0
+ */
+public class DumpFunctionExample {
+    public static void main(String[] args) throws Exception {
+        Replicator replicator = new RedisReplicator("redis://127.0.0.1:6379");
+        replicator.setRdbVisitor(new DumpRdbVisitor(replicator));
+        replicator.addEventListener(new EventListener() {
+            @Override
+            public void onEvent(Replicator replicator, Event event) {
+                if (event instanceof DumpFunction) {
+                    DumpFunction function = (DumpFunction) event;
+                    byte[] serialized = function.getSerialized();
+                    // your code goes here
+                    // you can use FUNCTION RESTORE to restore above serialized data to target redis
+                }
+            }
+        });
+        replicator.open();
+    }
+}