+++ src/main/java/com/moilioncircle/redis/replicator/rdb/dump/datatype/DumpFunction.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2016-2017 Leon Chen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.moilioncircle.redis.replicator.rdb.dump.datatype;
+
+import com.moilioncircle.redis.replicator.event.AbstractEvent;
+
+/**
+ * @author Leon Chen
+ * @since 3.6.0
+ * use FUNCTION RESTORE command can restore this serialized data to target redis
+ */
+public class DumpFunction extends AbstractEvent {
+    
+    private static final long serialVersionUID = 1L;
+    
+    private byte[] serialized;
+    
+    public byte[] getSerialized() {
+        return serialized;
+    }
+    
+    public void setSerialized(byte[] serialized) {
+        this.serialized = serialized;
+    }
+}